Changed 'Payment.invoice_id' field definition from PositiveIntegerField to BigIntegerField to support 'invoiceId' numbers that are currently in action on Yandex side.

+++ yandex_money/models.py
@@ -95,7 +95,7 @@ class Payment(models.Model):
     status = models.CharField(
         u'Статус', max_length=16, choices=STATUS.CHOICES,
         default=STATUS.PROCESSED)
-    invoice_id = models.PositiveIntegerField(
+    invoice_id = models.BigIntegerField(
         u'Номер транзакции оператора', blank=True, null=True)
     shop_amount = models.DecimalField(
         u'Сумма полученная на р/с', max_digits=15, decimal_places=2, blank=True,
